+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.10.1] - 2023-01-13
+
 ### Internal and API changes
 - Add ability to pass retrying executor in client builder ([#339](https://github.com/tarantool/cartridge-java/issues/339))
 - Do not use sync calls instead of async in retrying ([#299](https://github.com/tarantool/cartridge-java/pull/299))